Cast to pointer first
Fix: for Android Studio, unable to build with flag -Werror

+++ jni/src/GLES20Fix.c
@@ -6,9 +6,9 @@
 // ===========================================================
 
 void Java_org_andengine_opengl_GLES20Fix_glVertexAttribPointer (JNIEnv *env, jclass c, jint index, jint size, jint type, jboolean normalized, jint stride, jint offset) {
-	glVertexAttribPointer(index, size, type, normalized, stride, (void*) offset);
+	glVertexAttribPointer(index, size, type, normalized, stride, (void*) (intptr_t)offset);
 }
 
 void Java_org_andengine_opengl_GLES20Fix_glDrawElements (JNIEnv *env, jclass c, jint mode, jint count, jint type, jint offset) {
-	glDrawElements(mode, count, type, (void*) offset);
+	glDrawElements(mode, count, type, (void*) (intptr_t)offset);
 }
